langgraph_version/requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
openai==1.58.1
pydantic==2.8.2
langchain-openai==0.3.0
langgraph==0.2.50
requests==2.32.5
httpx==0.27.2
